Original punyforth code for wifi stepper
punyforth code created as a test to see whether wifi is practical for remote control of motors. Code runs on ESP12.

Found that UDP is just a few microseconds, and TCP is longer but not significantly considering how little a motor can move is the time between consecutive commands. So, yes, quite viable.